Refactor: moved string pre and post-processing to processors.py
